+++ oap-server/server-library/library-datacarrier-queue/src/main/java/org/apache/skywalking/oap/server/library/datacarrier/consumer/BulkConsumePool.java
@@ -78,7 +78,7 @@ public class BulkConsumePool implements ConsumerPool {
     @Override
     public void close(Channels channels) {
         for (MultipleChannelsConsumer consumer : allConsumers) {
-            consumer.shutdown();
+            consumer.close(channels);
         }
     }